+++ CHANGELOG.md
@@ -1,6 +1,15 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
+## [5.2.0] - 2017-01-??
+### Added:
+ - User *clearSearchHistory()* to remove search suggestions.
+ - Pins *tied()* returns the pinners who have tied this pin.
+ - Pins *searchInMyPins()* method.
+ - Inbox provider.
+### Removed:
+ - News and Conversations providers, their methods moved to Inbox provider.
+
 ## [5.1.7] - 2016-12-31
 ### Fixed:
  - Receiving confirmation email after registering business account.